moved controller code into own files
