+++ src/Controller.cpp
@@ -1,0 +1,6 @@
+#include "Controller.hpp"
+
+namespace trajectory_follower
+{
+Controller::~Controller(){}
+}
