+++ src/TrajectoryFollower.cpp
@@ -5,82 +5,6 @@
 
 using namespace trajectory_follower;
 
-Controller::~Controller(){}
-
-Motion2D& NoOrientationController::update(double speed, double distanceError, double angleError, double curvature, double variationOfCurvature)
-{
-    if (!configured)
-    {
-        throw std::runtime_error("controller is not configured.");
-    }
-
-    double u1, u2;
-    u1 = speed;
-    double direction = 1.;
-    if (speed < 0)
-    {
-        // Backward motion
-        direction = -1.;
-        u1 = fabs(u1);
-    }
-    // No orientation controller ( Page 806 ), Assuming k(d,theta_e)=K0*cos(theta_e)
-    u2 = -u1 * (tan(angleError) / l1 + distanceError * K0);
-
-    motionCommand.translation = u1 * direction;
-    motionCommand.rotation = u2;
-    return motionCommand;
-}
-
-Motion2D& ChainedController::update(double speed, double distanceError, double angleError, double curvature, double variationOfCurvature)
-{
-    if (!configured)
-    {
-        throw std::runtime_error("controller is not configured.");
-    }
-
-    double d_dot, s_dot, z2, z3, v1, v2, u1, u2;
-    u1 = speed;
-
-    double direction = 1.0;
-    if(speed < 0) {
-        // Backward motion
-        direction = -1.0;
-        u1 = fabs(u1);
-    }
-
-    d_dot = u1 * sin(angleError);
-    s_dot = u1 * cos(angleError) / (1.0-distanceError*curvature);
-
-    v1 = s_dot;
-    z2 = distanceError;
-    z3 = (1.0-(distanceError*curvature))*tan(angleError);
-
-    controllerIntegral += (s_dot * z2);
-    v2 = -fabs(v1)*K0*controllerIntegral - v1*K2*z2 - fabs(v1)*K3 * z3;
-    u2 = (v2 + (d_dot*curvature + distanceError*variationOfCurvature*s_dot)*tan(angleError))
-          /((1.0-distanceError*curvature)*(1+pow(tan(angleError),2))) + s_dot*curvature;
-
-    motionCommand.translation = u1*direction;
-    motionCommand.rotation = u2;
-    return motionCommand;
-}
-
-Motion2D& SamsonController::update(double speed, double distanceError, double angleError, double curvature, double variationOfCurvature)
-{
-    if (!configured)
-    {
-        throw std::runtime_error("controller is not configured.");
-    }
-
-    const double u1 = std::abs(speed);
-
-    const double u2 = -K2*distanceError*u1*(sin(angleError)/angleError) - K3*angleError;
-
-    motionCommand.translation = speed;
-    motionCommand.rotation = u2;
-    return motionCommand;
-}
-
 TrajectoryFollower::TrajectoryFollower()
     : configured(false),
       controllerType(CONTROLLER_UNKNOWN ),
